+++ views/homepage.handlebars
@@ -1,5 +1,5 @@
 <div class="sports">
-    {{#each sports as |sport|}}
+    {{#each sport}}
     {{> sport-cards}}
     {{/each}}
 </div>
