+++ views/layouts/main.handlebars
@@ -20,7 +20,7 @@
 
             <div class="col-sm">
               <!-- home/feed icon-->
-              <a href="/"  id="nav-link">Home</a>
+              <a href="/home"  id="nav-link">Home</a>
             </div>
 
             <div class="col-sm">
