+++ views/login.handlebars
@@ -14,11 +14,10 @@
             <button id="login-button" type="submit">Sign in</button>
         </div>
         <div>
-            <a href="/create-user" id="create-user-button">Create an accout</a>
+            <a href="/api/create-user" id="create-user-button">Create an accout</a>
         </div>
     </form>
   </div>
 </div>
 
 <script src="..public/js/login.js"></script>
-{{!-- <script src="./js/login.js"></script> --}}
