+++ views/partials/sport-cards.handlebars
@@ -1,6 +1,6 @@
 <section class="card">
-  <a href="/sport/{{id}}">
-    <img src="/images/{{sport.[0].filename}}" class="card-img-top" alt="image of the sport">
+  <a href="/api/sport/{{id}}">
+    <img src="../images/{{filename}}" class="card-img-top" alt="image of the sport">
     <div class="card-body">
       <h3 class="card-title">{{name}}</h3>
     </div>
