+++ views/profile.handlebars
@@ -1,4 +1,4 @@
-{{#if loggedIn}}
+{{!-- {{#if loggedIn}} --}}
 <section class ="profile">
     <div id="profile-header">
         <h4 id="profile-username">{{user.username}}'s Profile</h4>
@@ -16,10 +16,10 @@
     </div>
 
 <!-- Otherwise, provide a link to login -->
-{{else}}
+{{!-- {{else}}
 <a href="/login">Log in to see this page</a>
-{{/if}}
+{{/if}} --}}
 </section>
 
 
-<script src="/public/js/profile.js"></script>
+<script src="../js/createUser.js"></script>
